requirements.txt
certifi==2019.6.16
chardet==3.0.4
Django==2.1
django-bootstrap4==0.0.8
django-crispy-forms==1.7.2
django-filebrowser-no-grappelli==3.7.8
django-grappelli==2.12.3
django-tinymce4-lite==1.7.5
gunicorn==19.9.0
idna==2.8
jsmin==2.2.2
mysqlclient==1.4.2.post1
Pillow==6.1.0
pytz==2019.1
requests==2.22.0
sqlparse==0.3.0
stripe==2.32.1
urllib3==1.25.3
